Fix : Overflow issues in array_patterns

+++ c/array-patterns/array16_pattern.c
@@ -29,11 +29,11 @@ void __VERIFIER_assert(int cond) { if(!(cond)) { ERROR: __VERIFIER_error(); } }
 extern int __VERIFIER_nondet_int() ;
 extern short __VERIFIER_nondet_short() ;
 
-short ARR_SIZE ;
+signed long long ARR_SIZE ;
 
 int main()
 {
-	ARR_SIZE = __VERIFIER_nondet_short() ;
+	ARR_SIZE = (signed long long)__VERIFIER_nondet_short() ;
 	__VERIFIER_assume(ARR_SIZE > 0) ;
 
 	int array[ARR_SIZE][ARR_SIZE] ;
@@ -41,7 +41,7 @@ int main()
 	int row = 0, column = 0, num = -1 ;
 	signed long long sum = 0 ;
 	int temp ;
-	short index1,index2 ;
+	signed long long index1,index2 ;
 
 	for(row=0;row<ARR_SIZE;row++)
 		for(column=0;column<ARR_SIZE;column++)
@@ -50,8 +50,8 @@ int main()
 	while(1)
         {
 		
-		index1 = __VERIFIER_nondet_short() ;
-		index2 = __VERIFIER_nondet_short() ;
+		index1 = (signed long long)__VERIFIER_nondet_short() ;
+		index2 = (signed long long)__VERIFIER_nondet_short() ;
 		__VERIFIER_assume(index1>=0 && index1 < ARR_SIZE) ;
 		__VERIFIER_assume(index2>=0 && index2 < ARR_SIZE) ;
